Add creatures and characters
Closes #29 #49 #47.

Adds the ability to add Creatures and Characters to the current
map/campaign.

Characters are added to all maps in the campaign when created, and
creatures are only added to the current map.

Characters, by default, only have one instance on a map. When they are
dragged from the token drawer, they go onto the map and leave the
drawer.

Creatures are copied when they are dragged into the map, and each one is
given an identifier.

If you go to a map that doesn't have a creature you used on a previous
map, you can add it to the current map by going to "New Token" ->
"Library", there you see all characters and creatures that are used
anywhere in the campaign. You can edit them (which allows you to delete
them) and you can add them to the current map by clicking their token.

* rails-ujs is able to submit remote forms with file uploads. So
  creation and edit forms for tokens are now fully ajax. (#47)
* A `modal` layout has been extracted so views that are requested via
  ajax and want to be in a modal can render with that layout. In order to
  have an `html.erb` view, render inside of a `.js.erb` layout, you need
  to be explicit about the format in the controller. I've opted to do this
  with `respond_to :js` for conciseness, you can also use the block form
  of `respond_to` as well.