Fix token dragging in Firefox
Fixes #54

`clientX` and `clientY` are always set to 0 in Firefox on the drag event.

This is a bug in Firefox that has been open for 11 years
(https://bugzilla.mozilla.org/show_bug.cgi?id=505521).

We can get `clientX` and `clientY` in drag start, and by setting a
`dragover` handler on the entire map. To workaround this we set `clientX`
and `clientY` in the data of the map and use that instead of the
`clientX` and `clientY` from the `drag` event.