Fix displaying dm controls when switching maps
Fixes #55. Previously when switching maps it was broadcasting the
rendered map partial to all users, but that was being rendered by the dm
users.

This fixes the problem by broadcasting a chang_current_map operation to
everyone subscribed to the channel and having each person get their own
version of the map.

I explored a few different ways of handling this, and this was the
simplest/best fix I could figure.