Fix displaying dm controls when switching maps
Fixes #55. This addresses the issue in #55 by creating two channels, one
for the dm and the other for everyone else, and broadcasting map changes
with the appropriate html to each of them.

The change in the `_current_map` partial ended up being straightforward,
instead of checking within the partial if the person is the dm, move the
check up out of it and pass it in as a local variable.