Add support for token sizes
Closes #57. This adds support for multiple token sizes.

The duplication in the fields with go away, and we'll get the
introduction of sizes for characters with our upcoming token changes,
but I figured I'd plow through for now.

For now, we've made sure that the tokens look to be the right number of
squares on the maps we're using, but we understand that we ultimately
need to add #30.