Home page
First version of the home page. I've added a [Credits Wiki
Page](https://github.com/tableparty/table/wiki/Credits) for the new
artwork..

It was necessary to override the clearance behavior of redirecting to
the root after successful user creation or log in, as we now want to
redirect explicitly to the campaign index, as I want the home page to be
accessible to logged-in users.