Editing and deletion of maps
This adds the ability to edit and delete maps.

Map edits will be rebroadcast to all users. If you delete the current
map, it will be unset for everyone.

There is a foreign key constraint on `Campaign#current_map` that ensures
the current map cannot be deleted. Therefore, the Deletion Interactor
needs to check whether the map being deleted is the current map, and if
it is, then it unsets it and broadcasts that before deleting the map.