Improve sponsor code experience
Fixes #79. This change attempts to address the experience our user had
in #79 by presenting a meaningful error message when not providing a
sponsor code, or providing one that doesn't match.

The copy provided is based on what @jdbann said in the issue.

Instead of continuing to modify the controller, I took the opportunity
to extract this to Interactors.