Remove map and token jitter
Fixes #85

Previously, the token and map movement was relayed to everyone on the page,
including the person moving it, even though their browser had already
moved the item.

The movement was blocked while the token or map was being dragged, but
when it was released, rapid movements were likely to still be coming
over the cable.

This resolves the issue by having each person on the page ignore any of
their own map or token movements.

This is accomplished by having each person on the page generate a unique
"operator code". This code is sent along with the move map or token
operations. When operations are received on the channel, much like
ignoring the operation if beingDragged is true, the operation is ignored
if you were the one who did it.