Remove scaling from map coordinates
Previously, when the map was zoomed in or out, the x and y coordinates
were scaled to reflect the new zoom level. This meant that moving the
map required interaction with the server.

To allow players to move and zoom the map independently of the DM, it
would be easier if all calculation happened in the browser. So instead
we pass the raw x and y coordinates and adjust them to suit the zoom
level in the browser.