Handle all zooming on frontend
For zooming in or out, the actual zoom amount had to be retrieved from
the server. To allow moving and zooming the map without hitting the
backend, we need to provide the zoom scale to the frontend.

This change allows zooming in and out to just persist the new zoom level
rather than also passing an updated width, height and zoom amount.
Instead those calculations are handled by JS and CSS.