Collapse map selector
Fixes #70

This collapses the map selector down into a button based on @jdbann's
previous gifs.

Selecting a map closes the selector.

Along the way, it refactors the markup/style for the map controls to
make them more consistent and based on flexbox instead of absolute
positioning.

Finally, with the map selector collapsed, that was not ideal from a
blank-slate perspective, so if there is no current map selected, it will
automatically open the map selector to make it obvious how to create a
new map.