Closes #104, lays the groundwork for #84, and potentially closes #102
This PR introduces the ability to select individual or multiple tokens and then take actions on them. Right now the two supported actions are Delete and Stash. Stashing is the same as dragging the token back the drawer, deleting is a new action that we haven't had before, which completely deletes the token from the map. A confirmation dialog is presented before deleting since it's a destructive action, but not for stashing.

Other notes:
* Hold shift (or command on Mac, control on Windows) while clicking to select multiple tokens
* The action buttons are enabled and disabled as you select or unselect tokens.
* It is possible to select and delete tokens that are in the drawer
* It is possible to delete a character from the map. Doing so only deletes it from this map, it will still be in the campaign in the library and on other maps, including automatically added to newly created maps.

To round out the behavior all of this functionality is keyboard accessible:

* Left arrow and right arrow are used to move the selected token
* Hold shift (or command on Mac, control on Windows) while pressing right arrow or left arrow to select multiple tokens using the keyboard
* When tokens are selected, pressing Delete or Backspace will delete them (after confirmation)
* `s` key short stashes the tokens