Individual token editing
Builds on #113 to introduce the ability to edit the details of an individual token and then have that broadcast to all players. You can edit the name, identifier, size, and the image of the token. The keyboard short for this is `e`.

This also potentially helps with #99, allowing DMs to swap out the image on a token during play, though you'd need to keep the files handy.

When you edit the token, you're editing that individual instance of the token, and it doesn't affect the other tokens on the map or campaign. When a token is from a template (which is all tokens) we display a link to edit the template from this screen.

Changing the template is not something that is currently broadcast.